Move shipclass and cockpit display classes into separate files
