+++ code/scripting/api/shipclass.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "scripting/ade_api.h"
+
+namespace scripting {
+namespace api {
+
+DECLARE_ADE_OBJ(l_Shipclass, int);
+
+}
+}
